move lists api to separate class
