+++ automated_api.py
@@ -34,6 +34,7 @@ from ayon_api.server_api import (  # noqa: E402
     ServerAPI,
     _PLACEHOLDER,
     _ActionsAPI,
+    _ListsAPI,
 )
 from ayon_api.utils import NOT_SET  # noqa: E402
 
@@ -294,6 +295,7 @@ def prepare_api_functions(api_globals):
     functions = []
     _items = list(ServerAPI.__dict__.items())
     _items.extend(_ActionsAPI.__dict__.items())
+    _items.extend(_ListsAPI.__dict__.items())
     for attr_name, attr in _items:
         if (
             attr_name.startswith("_")
